Workaround for Amazon inconsistency no longer required
Amazon removed the inconsistency and broke the code that tried to work around said inconsistency. Of course had the workaround not be applied it would have not been able to work in the previous inconsistent state. So no matter what we do, Amazon's inconsistency leads to borken code either now or into the future.

(╯°□°)╯︵ ┻━┻