Integration tests (#14)
* Mini test suite

This is essentially an integration test. I am not doing unit
tests because my experience is that what worked before or
implemented per the documentation may stop working because
Amazon changed something on their end.

These tests are far slower to execute but it's not like this
library gets any crazy amounts of development action anymore.
Also, I am more interested in making sure I can actually use
this in the real world where what matter is whether it works,
not if it conforms to a unit test suite. I mean, unit tests
would be great... if Amazon's documentation was guaranteed to
be 100% correct and read only in one, authoritative way. Right?