# [HIGH] Cannot do multipart uploads to Ceph S3
The problem was that the UploadId returned by Ceph S3
contains a tilde (`~`). The v4 signature was using
`urlencode()` replacing `+` with `%20` to emulate RFC 3986
encoding. This was wrong, because `urlencode()`, despite
what PHP documents, differs in one more way: it
URL-encodes tildes, whereas RFC 3986 doesn't. Therefore,
the correct function to use is `rawurlencode()`.

The funny thing is that if you read the documentation
carefully, there is indeed one character difference
in the description of the two encoding functions. The
`rawurlencode()` one does say it does not encode `-_.~`
whereas `urlencode()` says it does not encode `-_.`.

A bug caused by a single character. Such a classic!